% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks_plot.R
\name{cooks_plot}
\alias{cooks_plot}
\title{Cook's Distance Plot}
\usage{
cooks_plot(model, label)
}
\arguments{
\item{model}{a glm or lm model}

\item{label}{text labels for plot}
}
\value{
a cook's distance plot
}
\description{
Cook's Distance Plot
}
