% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfbeta_plot.R
\name{dfbeta_plot}
\alias{dfbeta_plot}
\title{dfbeta Plot}
\usage{
dfbeta_plot(model, label)
}
\arguments{
\item{model}{a glm or lm model}

\item{label}{text labels for plot}
}
\value{
a Dan-Gillen-Style dfbeta plot
}
\description{
dfbeta Plot
}
