% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cov.R
\name{empirical_cov}
\alias{empirical_cov}
\title{empirical_cov}
\usage{
empirical_cov(x, y, id, m)
}
\arguments{
\item{x}{x vector}

\item{y}{y vector}

\item{id}{id vector}

\item{m}{number of breaks}
}
\value{
list of rmat, cmat, cormat, nmat
}
\description{
empirical_cov
}
