% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_lda_variogram.R
\name{gg_lda_variogram}
\alias{gg_lda_variogram}
\title{gg_lda_variogram}
\usage{
gg_lda_variogram(x, y, id, title = NULL)
}
\arguments{
\item{x}{x vector}

\item{y}{y vector}

\item{id}{id vector}

\item{title}{title of plot}
}
\value{
Dan Gillen style variogram with ggplpot
}
\description{
Produce Dan Gillen style variograms with ggplot
}
