% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_matrix.R
\name{print_matrix}
\alias{print_matrix}
\title{print_matrix}
\usage{
print_matrix(x, digits = 2, mat.type = "b")
}
\arguments{
\item{x}{a matrix}

\item{digits}{the number of digits to round the matrix}

\item{mat.type}{"", "b", "B", "p", "v", "V", or "small"}
}
\value{
LaTeX code for printed matrix
}
\description{
print_matrix
}
