% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vector.R
\name{print_vector}
\alias{print_vector}
\title{The length of a string (in characters).}
\usage{
print_vector(x, oxford.comma = T)
}
\arguments{
\item{x}{input character vector}
}
\value{
the input string formatted as in an english sentence
}
\description{
The length of a string (in characters).
}
\examples{
print_vector(letters)
print_vector(letters, oxford.comma = F)
}
