% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_sum.R
\name{robust_sum}
\alias{robust_sum}
\title{robust glm summary}
\usage{
robust_sum(model, vcov = sandwich::vcovHC(model, type = "HC3"),
  level = 0.95, transform = T)
}
\arguments{
\item{model}{glm object}

\item{vcov}{robust covariance matrix}

\item{transform}{logical}
}
\value{
data frame with robust summary
}
\description{
robust glm summary
}
